refactor: create actions dir to host actions
