+++ src/binary-window/actions/close.js
@@ -1,4 +1,4 @@
-import { getSashIdFromPane } from '../frame/frame-utils';
+import { getSashIdFromPane } from '../../frame/frame-utils';
 
 export default {
   label: '',
