+++ src/binary-window/actions/index.js
@@ -1,0 +1,1 @@
+export { default, BUILTIN_ACTIONS } from './module';
